chore(metadata-manager): Update vuedraggable, vue, vue-template-compiler. Add test utils + mocha rep
Vuedraggable was updated to make sure the bug wasn't there. Updated to 2.18.1, because newer
versions had issues (2.19.0 gives error, 2.19.1 wont drag, 2.20 doesn't do anything). Test utils
were added to write test for vue component. Mocha reporter was added to make it easier to debug
tests. Vue was updated because the version was not compatible with the vue test utils and the the
vue-template-compiler had to be the same version as vue.