fix(one-click-importer): Set height of file input to 100%
Set height of file input in one click importer to 100% to override the faulty bootstrap theme
calculation. We chose to keep the current file input rather than replace it by the bootstrap one
because we like the browser translations, those are lost when the bootstrap variant is used.

Fixes molgenis/molgenis#8368