chore(core-ui): cache proxy config for max 10m (#128)
The unpkg resources with specific versions have long expiry dates.
If we get there using a fully versioned URL that is correct.

If we got there because we're following unpkg redirects it is incorrect.
The redirects may change.

If we get there because we've rewritten directly to a specific version
it is also incorrect. The proxy config may change.